refactor: move transcription and formatting utilities to utils/
- Move transcription-segments.ts to utils/transcription.ts
- Move formatReasoning to utils/format.ts and delete standalone file
- Update all imports to use new paths
