+++ src/components/batch/batch-transcription-view.tsx
@@ -51,7 +51,7 @@ import { BatchTranscriptionFileList } from "./batch-transcription-file-list"
 import { useBatchSelection } from "@/hooks/batch/use-batch-selection"
 import { CopyTranscriptionSettingsDialog } from "./copy-transcription-settings-dialog"
 import { useWhisperSettingsStore } from "@/stores/settings/use-whisper-settings-store"
-import { generateWordsSubtitles, generateSegmentsTranscription } from "@/lib/transcription-segments"
+import { generateWordsSubtitles, generateSegmentsTranscription } from "@/lib/utils/transcription"
 import { parseSubtitle } from "@/lib/subtitles/parse-subtitle"
 
 interface BatchTranscriptionViewProps {
