+++ src/components/transcribe/transcription-main.tsx
@@ -26,7 +26,7 @@ import { SubtitleTranslated } from "@/types/subtitles"
 import { useRouter } from "next/navigation"
 import { cn, createUtf8SubtitleBlob } from "@/lib/utils"
 import { useLocalSettingsStore } from "@/stores/settings/use-local-settings-store"
-import { generateWordsSubtitles, generateSegmentsTranscription } from "@/lib/transcription-segments"
+import { generateWordsSubtitles, generateSegmentsTranscription } from "@/lib/utils/transcription"
 import { useTranscriptionHandler } from "@/hooks/handler/use-transcription-handler"
 import { useWhisperSettingsStore } from "@/stores/settings/use-whisper-settings-store"
 import { WhisperSettingsPanel } from "./whisper-settings-panel"
