+++ src/hooks/batch/use-batch-transcription-handler.ts
@@ -11,7 +11,7 @@ import { toast } from "sonner"
 import { uploadFile } from "@/lib/api/file-upload"
 import { useQueryClient, useQuery } from "@tanstack/react-query"
 import { parseSubtitle } from "@/lib/subtitles/parse-subtitle"
-import { generateWordsSubtitles, generateSegmentsTranscription } from "@/lib/transcription-segments"
+import { generateWordsSubtitles, generateSegmentsTranscription } from "@/lib/utils/transcription"
 import { BatchFile } from "@/types/batch"
 import { useClientIdStore } from "@/stores/ui/use-client-id-store"
 import { useUnsavedChanges } from "@/contexts/unsaved-changes-context"
