+++ src/lib/utils/format.ts
@@ -20,3 +20,14 @@ export function formatTokens(tokens: number | undefined): string {
   }
   return `${tokens} tokens`
 }
+
+export function formatReasoning(text: string): string {
+  let isFirst = true
+  return text.replace(/(\*\*[^*]+\*\*)\n\n/g, (match) => {
+    if (isFirst) {
+      isFirst = false
+      return match
+    }
+    return '\n\n' + match
+  })
+}
